+++ app/src/main/java/org/sopt/certi/domain/model/comment/CommentData.kt
@@ -13,10 +13,9 @@ data class CommentData(
     val createdTime: String, // 생성일자
     val lastModifiedTime: String, // 수정일자
     val isLike: Boolean, // 댓글을 조회하는 사용자가 해당 댓글에 좋아요를 눌렀는지
+    val likeCount: Int, // 좋아요 갯수
+
     val totalPages: Int, // Integer	댓글 총 페이지
     val totalElements: Int, // Integer	댓글 요소
-    val isLast: Boolean, // Boolean	마지막 페이지인지
-
-    // 아래는 필요하지만 문서에 없는것들
-    val likeCount: Int
+    val isLast: Boolean // Boolean	마지막 페이지인지
 )
