+++ app/src/main/res/values/strings.xml
@@ -222,5 +222,6 @@
     <string name="comment_delete_button">삭제</string>
     <string name="comment_like_count">좋아요 %s</string>
     <string name="comment_report">신고</string>
+    <string name="comment_user_info">(%s, %s)</string>
 
 </resources>
